fix(Bug): Resolve "$viewModel" not found error in ParkingDetailsView
This commit resolves the "Cannot find  in scope" error in `ParkingDetailsView.swift`.

The error occurred because `viewModel` was declared as a `let` constant, preventing the creation of bindings to its properties.

This commit fixes the issue by changing `viewModel` to an `@ObservedObject`, allowing its properties to be passed as bindings to subviews. The custom `init` method, which is no longer needed with `@ObservedObject`, has also been removed.